+++ coxdev/base.py
@@ -1,419 +1,403 @@
-# these should be done in C ideally
+"""
+Cox Proportional Hazards Model Deviance Computation.
+
+This module provides efficient computation of Cox model deviance, gradients,
+and Hessian information matrices for survival analysis with support for
+different tie-breaking methods (Efron and Breslow).
+"""
+
+from dataclasses import dataclass, InitVar
+from typing import Literal, Optional
+# for Hessian
+
+from scipy.sparse.linalg import LinearOperator
+
+from . import _version
+__version__ = _version.get_versions()['version']
 
 import numpy as np
+from joblib import hash as _hash
 
-def _forward_cumsum(sequence, output):
-    '''
-    compute cumsum with a padding of 0 at the beginning
-    '''
-    output[:] = np.cumsum(np.hstack([0, sequence]))
+from .coxc import (cox_dev as _cox_dev,
+                   hessian_matvec as _hessian_matvec,
+                   compute_sat_loglik as _compute_sat_loglik,
+                   c_preprocess)
 
-def _reverse_cumsums(sequence,
-                     event_buffer,
-                     start_buffer,
-                     event_order,
-                     start_order,
-                     do_event=False,
-                     do_start=False):
+    
+@dataclass
+class CoxDevianceResult(object):
     """
-    Compute reversed cumsums of a sequence
-    in start and / or event order with a 0 padded at the end.
+    Result object containing Cox model deviance computation results.
+    
+    Attributes
+    ----------
+    linear_predictor : np.ndarray
+        The linear predictor values (X @ beta) used in the computation.
+    sample_weight : np.ndarray
+        Sample weights used in the computation.
+    loglik_sat : float
+        Saturated log-likelihood value.
+    deviance : float
+        Computed deviance value.
+    gradient : Optional[np.ndarray]
+        Gradient of the deviance with respect to the linear predictor.
+    diag_hessian : Optional[np.ndarray]
+        Diagonal of the Hessian matrix.
+    __hash_args__ : str
+        Hash string for caching results.
+    """
+
+    linear_predictor: np.ndarray
+    sample_weight: np.ndarray
+    loglik_sat: float
+    deviance: float
+    gradient: Optional[np.ndarray]
+    diag_hessian: Optional[np.ndarray]
+    __hash_args__: str
+
+
+@dataclass
+class CoxDeviance(object):
+    """
+    Cox Proportional Hazards Model Deviance Calculator.
+    
+    This class provides efficient computation of Cox model deviance, gradients,
+    and Hessian information matrices. It supports both Efron and Breslow
+    tie-breaking methods and handles left-truncated survival data.
+    
+    Parameters
+    ----------
+    event : np.ndarray
+        Event times (failure times) for each observation.
+    status : np.ndarray
+        Event indicators (1 for event occurred, 0 for censored).
+    start : np.ndarray, optional
+        Start times for left-truncated data. If None, assumes no truncation.
+    tie_breaking : {'efron', 'breslow'}, default='efron'
+        Method for handling tied event times.
+        
+    Attributes
+    ----------
+    tie_breaking : str
+        The tie-breaking method being used.
+    _have_start_times : bool
+        Whether start times are provided.
+    _efron : bool
+        Whether Efron's method is being used for tie-breaking.
     """
     
-    # pad by 1 at the end length=n+1 for when last=n-1    
-
-    if do_event:
-        seq_event = np.hstack([sequence[event_order], 0])
-        event_buffer[:] = np.cumsum(seq_event[::-1])[::-1]
-
-    if do_start:
-        seq_start = np.hstack([sequence[start_order], 0])
-        start_buffer[:] = np.cumsum(seq_start[::-1])[::-1]  # length=n+1
-
-def _to_native_from_event(arg,
-                          event_order,
-                          reorder_buffer):
-    """
-    reorder an event-ordered vector into native order,
-    uses forward_scratch_buffer to make a temporary copy
-    """
-    reorder_buffer[:] = arg
-    arg[event_order] = reorder_buffer
-
-def _to_event_from_native(arg,
-                          event_order,
-                          reorder_buffer):
-    """
-    reorder an event-ordered vector into native order,
-    uses forward_scratch_buffer to make a temporary copy
-    """
-    reorder_buffer[:] = arg[event_order]
-
-def _forward_prework(status,
-                     w_avg,
-                     scaling,
-                     risk_sums,
-                     i,
-                     j,
-                     moment_buffer,
-                     use_w_avg=True,
-                     arg=None):
-    """
-    we need some sort of cumsums of scaling**i / risk_sums**j weighted by w_avg (within status==1)
-
-    this function fills in appropriate buffer
-    """
-    if use_w_avg:
-        moment_buffer[:] = status * w_avg * (scaling**i) / (risk_sums**j)
-    else:
-        moment_buffer[:] = status * (scaling**i) / (risk_sums**j)
-    if arg is not None:
-        moment_buffer[:] *= arg
-
-def _compute_sat_loglik(_first,
-                        _last,
-                        _weight, # in natural order!!!
-                        _event_order,
-                        _status,
-                        W_status):
+    event: InitVar[np.ndarray]
+    status: InitVar[np.ndarray]
+    start: InitVar[np.ndarray]=None
+    tie_breaking: Literal['efron', 'breslow'] = 'efron'
     
-    _forward_cumsum(_weight[_event_order] * _status, W_status)
-    sums = W_status[_last+1] - W_status[_first]
-    loglik_sat = 0
-    prev_first = -1
-    for f, s in zip(_first, sums):
-        if s > 0 and f != prev_first:
-            loglik_sat -= s * np.log(s)
-        prev_first = f
+    def __post_init__(self,
+                      event,
+                      status,
+                      start=None):
+        """
+        Initialize the CoxDeviance object with survival data.
+        
+        Parameters
+        ----------
+        event : np.ndarray
+            Event times for each observation.
+        status : np.ndarray
+            Event indicators (1 for event, 0 for censored).
+        start : np.ndarray, optional
+            Start times for left-truncated data.
+        """
+        event = np.asarray(event)
 
-    return loglik_sat
+        status_arr = np.asarray(status)
+        if not set(np.unique(status_arr)).issubset(set([0,1])):
+            raise ValueError('status must be binary')
+        status = status_arr.astype(np.int32)
+        
+        nevent = event.shape[0]
 
-# Core function: compute log-likelihood
-
-def _cox_dev(eta,           # eta is in native order  -- assumes centered (or otherwise normalized for numeric stability)
-             sample_weight, # sample_weight is in native order
-             exp_w,
-             event_order,   
-             start_order,
-             status,        # everything below in event order
-             first,
-             last,
-             scaling,
-             event_map,
-             start_map,
-             loglik_sat,
-             T_1_term,
-             T_2_term,
-             grad_buffer,
-             diag_hessian_buffer,
-             diag_part_buffer,
-             w_avg_buffer,
-             event_reorder_buffers,
-             risk_sum_buffers,
-             forward_cumsum_buffers,
-             forward_scratch_buffer,
-             reverse_cumsum_buffers,
-             have_start_times=True,
-             efron=False):
-
-    n = eta.shape[0]
-    
-    _to_event_from_native(eta, event_order, event_reorder_buffers[0])
-    eta_event = event_reorder_buffers[0]
-
-    _to_event_from_native(sample_weight, event_order, event_reorder_buffers[1])
-    w_event = event_reorder_buffers[1]
-
-    _to_event_from_native(exp_w, event_order, event_reorder_buffers[2])
-    exp_eta_w_event = event_reorder_buffers[2]
-   
-    if have_start_times:
-        _sum_over_risk_set(exp_w, # native order
-                           event_order,
-                           start_order,
-                           first,
-                           last,
-                           event_map,
-                           scaling,
-                           efron,
-                           risk_sum_buffers[0],
-                           reverse_cumsum_buffers[:2])
-    else:
-        _sum_over_risk_set(exp_w, # native order
-                           event_order,
-                           start_order,
-                           first,
-                           last,
-                           None,
-                           scaling,
-                           efron,
-                           risk_sum_buffers[0],
-                           reverse_cumsum_buffers[:2])
-
-    event_cumsum = reverse_cumsum_buffers[0]
-    start_cumsum = reverse_cumsum_buffers[1]
-    risk_sums = risk_sum_buffers[0]
-    
-    # some ordered terms to complete likelihood
-    # calculation
-
-    # w_cumsum is only used here, can write over forward_cumsum_buffers
-    # after computing w_avg
-    _forward_cumsum(w_event, forward_cumsum_buffers[0])
-    w_cumsum = forward_cumsum_buffers[0]
-    w_avg_buffer[:] = ((w_cumsum[last + 1] - w_cumsum[first]) /
-                       (last + 1 - first))
-    w_avg = w_avg_buffer # shorthand
-
-    loglik = ((w_event * eta_event * status).sum() -
-              np.sum(np.log(np.array(risk_sums)) * w_avg * status))
-
-    # forward cumsums for gradient and Hessian
-    
-    # length of cumsums is n+1
-    # 0 is prepended for first(k)-1, start(k)-1 lookups
-    # a 1 is added to all indices
-
-    _forward_prework(status, w_avg, scaling, risk_sums, 0, 1, forward_scratch_buffer)
-    A_01 = forward_scratch_buffer
-    _forward_cumsum(A_01, forward_cumsum_buffers[0]) # length=n+1 
-    C_01 = forward_cumsum_buffers[0]
-
-    _forward_prework(status, w_avg, scaling, risk_sums, 0, 2, forward_scratch_buffer)
-    A_02 = forward_scratch_buffer
-    _forward_cumsum(A_02, forward_cumsum_buffers[1]) # length=n+1
-    C_02 = forward_cumsum_buffers[1]
-    
-    if not efron:
-        if have_start_times:
-
-            # +1 for start_map? depends on how  
-            # a tie between a start time and an event time
-            # if that means the start individual is excluded
-            # we should add +1, otherwise there should be
-            # no +1 in the [start_map+1] above
-
-            T_1_term[:] = C_01[last+1] - C_01[start_map]
-            T_2_term[:] = C_02[last+1] - C_02[start_map]
-
+        if start is None:
+            start = -np.ones(nevent) * np.inf
+            self._have_start_times = False
         else:
-            T_1_term[:] = C_01[last+1]
-            T_2_term[:] = C_02[last+1]
-    else:
-        # compute the other necessary cumsums
+            start = np.asarray(start)
+            self._have_start_times = True
+
+        (self._preproc,
+         self._event_order,
+         self._start_order) = c_preprocess(start,
+                                           event,
+                                           status)
+        self._event_order = self._event_order.astype(np.int32)
+        self._start_order = self._start_order.astype(np.int32)
         
-        _forward_prework(status, w_avg, scaling, risk_sums, 1, 1, forward_scratch_buffer)
-        A_11 = forward_scratch_buffer
-        _forward_cumsum(A_11, forward_cumsum_buffers[2]) # length=n+1
-        C_11 = forward_cumsum_buffers[2]
+        self._efron = self.tie_breaking == 'efron' and np.linalg.norm(self._preproc['scaling']) > 0
 
-        _forward_prework(status, w_avg, scaling, risk_sums, 2, 1, forward_scratch_buffer)
-        A_21 = forward_scratch_buffer
-        _forward_cumsum(A_21, forward_cumsum_buffers[3]) # length=n+1
-        C_21 = forward_cumsum_buffers[3]
-
-        _forward_prework(status, w_avg, scaling, risk_sums, 2, 2, forward_scratch_buffer)
-        A_22 = forward_scratch_buffer
-        _forward_cumsum(A_22, forward_cumsum_buffers[4]) # length=n+1
-        C_22 = forward_cumsum_buffers[4]
-
-        T_1_term[:] = (C_01[last+1] - 
-                       (C_11[last+1] - C_11[first]))
-        T_2_term[:] = ((C_22[last+1] - C_22[first]) 
-                       - 2 * (C_21[last+1] - C_21[first]) + 
-                       C_02[last+1])
-
-        if have_start_times:
-            T_1_term -= C_01[start_map]
-            T_2_term -= C_02[first]
-    
-    # could do multiply by exp_w after reorder...
-    # save a reorder of w * exp(eta)
-
-    diag_part_buffer[:] = exp_eta_w_event * T_1_term
-    grad_buffer[:] = w_event * status - diag_part_buffer
-    grad_buffer *= -2
-    # now the diagonal of the Hessian
-
-    diag_hessian_buffer[:] = exp_eta_w_event**2 * T_2_term - diag_part_buffer
-    diag_hessian_buffer *= -2
-    
-    _to_native_from_event(grad_buffer, event_order, forward_scratch_buffer)
-    _to_native_from_event(diag_hessian_buffer, event_order, forward_scratch_buffer)
-    _to_native_from_event(diag_part_buffer, event_order, forward_scratch_buffer)
-    
-    deviance = 2 * (loglik_sat - loglik)
-
-    return deviance
-
-def _sum_over_events(event_order,
-                     start_order,
-                     first,
-                     last,
-                     start_map,
-                     scaling,
-                     status,
-                     efron,
-                     forward_cumsum_buffers,
-                     forward_scratch_buffer,
-                     value_buffer):
-    '''
-    compute sum_i (d_i Z_i ((1_{t_k>=t_i} - 1_{s_k>=t_i}) - sigma_i (1_{i <= last(k)} - 1_{i <= first(k)-1})
-    '''
+        self._status = np.asarray(self._preproc['status'])
+        self._event = np.asarray(self._preproc['event'])
+        self._start = np.asarray(self._preproc['start'])
+        self._first = np.asarray(self._preproc['first']).astype(np.int32)
+        self._last = np.asarray(self._preproc['last']).astype(np.int32)
+        self._scaling = np.asarray(self._preproc['scaling'])
+        self._event_map = np.asarray(self._preproc['event_map']).astype(np.int32)
+        self._start_map = np.asarray(self._preproc['start_map']).astype(np.int32)
+        self._first_start = self._first[self._start_map]
         
-    have_start_times = start_map is not None
+        if not np.all(self._first_start == self._start_map):
+            raise ValueError('first_start disagrees with start_map')
 
-    n = status.shape[0]
+        n = self._status.shape[0]
 
-    _forward_cumsum(forward_scratch_buffer, forward_cumsum_buffers[0]) # length=n+1
-    C_arg = forward_cumsum_buffers[0]
-    
-    value_buffer[:] = C_arg[last+1]
-    if have_start_times:
-        value_buffer -= C_arg[start_map]
-
-    if efron:
-        forward_scratch_buffer[:] *= scaling
-        _forward_cumsum(forward_scratch_buffer, forward_cumsum_buffers[1]) # length=n+1
-        C_arg_scale = forward_cumsum_buffers[1]
-        value_buffer -= C_arg_scale[last+1] - C_arg_scale[first]
-
-def _sum_over_risk_set(arg,
-                       event_order,
-                       start_order,
-                       first,
-                       last,
-                       event_map,
-                       scaling,
-                       efron,
-                       risk_sum_buffer,
-                       reverse_cumsum_buffers):
-
-    '''
-    arg is in native order
-    returns a sum in event order
-    '''
-
-    have_start_times = event_map is not None
-
-    event_cumsum = reverse_cumsum_buffers[0]
-    start_cumsum = reverse_cumsum_buffers[1]
-    
-    _reverse_cumsums(arg,
-                     event_cumsum,
-                     start_cumsum,
-                     event_order,
-                     start_order,
-                     do_event=True,
-                     do_start=have_start_times)
-
-    if have_start_times:
-        risk_sum_buffer[:] = event_cumsum[first] - start_cumsum[event_map]
-    else:
-        risk_sum_buffer[:] = event_cumsum[first]
+        # allocate necessary memory
         
-    # compute the Efron correction, adjusting risk_sum if necessary
+        self._T_1_term = np.zeros(n)
+        self._T_2_term = np.zeros(n)
+        self._event_reorder_buffers = [np.zeros(n) for i in range(3)]
+        self._forward_cumsum_buffers = [np.zeros(n+1) for i in range(5)]
+        self._forward_scratch_buffer = np.zeros(n)
+        self._reverse_cumsum_buffers = [np.zeros(n+1) for i in range(4)]
+        self._risk_sum_buffers = [np.zeros(n) for i in range(2)]
+        self._hess_matvec_buffer = np.zeros(n)
+        self._grad_buffer = np.zeros(n)
+        self._diag_hessian_buffer = np.zeros(n)
+        self._diag_part_buffer = np.zeros(n)
+        self._w_avg_buffer = np.zeros(n)
+        self._exp_w_buffer = np.zeros(n)
+
+    def __call__(self,
+                 linear_predictor,
+                 sample_weight=None):
+        """
+        Compute Cox model deviance and related quantities.
+        
+        Parameters
+        ----------
+        linear_predictor : np.ndarray
+            Linear predictor values (X @ beta).
+        sample_weight : np.ndarray, optional
+            Sample weights. If None, uses equal weights.
+            
+        Returns
+        -------
+        CoxDevianceResult
+            Object containing deviance, gradient, and Hessian diagonal.
+        """
+        if sample_weight is None:
+            sample_weight = np.ones_like(linear_predictor)
+        else:
+            sample_weight = np.asarray(sample_weight)
+
+        linear_predictor = np.asarray(linear_predictor)
+            
+        cur_hash = _hash([linear_predictor, sample_weight])
+        if not hasattr(self, "_result") or self._result.__hash_args__ != cur_hash:
+
+            loglik_sat = _compute_sat_loglik(self._first,
+                                             self._last,
+                                             sample_weight, # in natural order
+                                             self._event_order,
+                                             self._status,
+                                             self._forward_cumsum_buffers[0])
+            
+            eta = np.asarray(linear_predictor)
+            sample_weight = np.asarray(sample_weight)
+            eta = eta - eta.mean()
+            self._exp_w_buffer[:] = sample_weight * np.exp(np.clip(eta, -np.inf, 30))
+
+
+
+
+            deviance = _cox_dev(eta,
+                                sample_weight,
+                                self._exp_w_buffer,
+                                self._event_order,
+                                self._start_order,
+                                self._status,
+                                self._first,
+                                self._last,
+                                self._scaling,
+                                self._event_map,
+                                self._start_map,
+                                loglik_sat,
+                                self._T_1_term,
+                                self._T_2_term,
+                                self._grad_buffer,
+                                self._diag_hessian_buffer,
+                                self._diag_part_buffer,
+                                self._w_avg_buffer,
+                                self._event_reorder_buffers,
+                                self._risk_sum_buffers, #[0] is for coxdev, [1] is for hessian...
+                                self._forward_cumsum_buffers,
+                                self._forward_scratch_buffer,
+                                self._reverse_cumsum_buffers, #[0:2] are for risk sums, [2:4] used for hessian risk*arg sums
+                                self._have_start_times,
+                                self._efron)
+                                
+            # shorthand, for reference in hessian_matvec
+            self._event_cumsum = self._reverse_cumsum_buffers[0]
+            self._start_cumsum = self._reverse_cumsum_buffers[1]
+
+            self._result = CoxDevianceResult(linear_predictor=linear_predictor,
+                                             sample_weight=sample_weight,
+                                             loglik_sat=loglik_sat,
+                                             deviance=deviance,
+                                             gradient=self._grad_buffer.copy(),
+                                             diag_hessian=self._diag_hessian_buffer.copy(),
+                                             __hash_args__=cur_hash)
+            
+        return self._result
+
+    def information(self,
+                    linear_predictor,
+                    sample_weight=None):
+        """
+        Compute the information matrix (negative Hessian) as a linear operator.
+        
+        Parameters
+        ----------
+        linear_predictor : np.ndarray
+            Linear predictor values (X @ beta).
+        sample_weight : np.ndarray, optional
+            Sample weights. If None, uses equal weights.
+            
+        Returns
+        -------
+        CoxInformation
+            Linear operator representing the information matrix.
+        """
+        result = self(linear_predictor,
+                      sample_weight)
+        return CoxInformation(result=result,
+                              coxdev=self)
+
+@dataclass
+class CoxInformation(LinearOperator):
+    """
+    Linear operator representing the Cox model information matrix.
     
-    if efron:
-        # for K events,
-        # this results in risk sums event_cumsum[first] to
-        # event_cumsum[first] -
-        # (K-1)/K [event_cumsum[last+1] - event_cumsum[first]
-        # or event_cumsum[last+1] + 1/K [event_cumsum[first] - event_cumsum[last+1]]
-        # to event[cumsum_first]
-        delta = (event_cumsum[first] - 
-                 event_cumsum[last+1])
-        risk_sum_buffer[:] -= delta * scaling
-
-def _hessian_matvec(arg,           # arg is in native order
-                    eta,           # eta is in native order 
-                    sample_weight, # sample_weight is in native order
-                    risk_sums,
-                    diag_part,
-                    w_avg,
-                    exp_w,
-                    event_cumsum,
-                    start_cumsum,
-                    event_order,   
-                    start_order,
-                    status,        # everything below in event order
-                    first,
-                    last,
-                    scaling,
-                    event_map,
-                    start_map,
-                    risk_sum_buffers,
-                    forward_cumsum_buffers,
-                    forward_scratch_buffer,
-                    reverse_cumsum_buffers,
-                    hess_matvec_buffer,
-                    have_start_times=True,
-                    efron=False):                    
-
-    #breakpoint()
-    if have_start_times:
-        # now in event_order
-        _sum_over_risk_set(exp_w * arg, # in native order
-                           event_order,
-                           start_order,
-                           first,
-                           last,
-                           event_map,
-                           scaling,
-                           efron,
-                           risk_sum_buffers[1], 
-                           reverse_cumsum_buffers[2:4])
-    else:
-        _sum_over_risk_set(exp_w * arg, # in native order
-                           event_order,
-                           start_order,
-                           first,
-                           last,
-                           None,
-                           scaling,
-                           efron,
-                           risk_sum_buffers[1], 
-                           reverse_cumsum_buffers[2:4])
-
-    risk_sums_arg = risk_sum_buffers[1]
-
-    # E_arg = risk_sums_arg / risk_sums -- expecations under the probabilistic interpretation
-    # forward_scratch_buffer[:] = status * w_avg * E_arg / risk_sums
-
-    # one less step to compute from above representation
-    forward_scratch_buffer[:] = status * w_avg * risk_sums_arg / risk_sums**2
-    print(f'forward_scratch_buffer {forward_scratch_buffer}')
+    This class provides matrix-vector multiplication with the information
+    matrix (negative Hessian) of the Cox model, allowing efficient computation
+    without explicitly forming the full matrix.
     
-    if have_start_times:
-        _sum_over_events(event_order,
-                         start_order,
-                         first,
-                         last,
-                         start_map,
-                         scaling,
-                         status,
-                         efron,
-                         forward_cumsum_buffers,
-                         forward_scratch_buffer,
-                         hess_matvec_buffer)
-    else:
-        _sum_over_events(event_order,
-                         start_order,
-                         first,
-                         last,
-                         None,
-                         scaling,
-                         status,
-                         efron,
-                         forward_cumsum_buffers,
-                         forward_scratch_buffer,
-                         hess_matvec_buffer)
-    print(f'hess_matvec_buffer {hess_matvec_buffer}')        
-    
-    _to_native_from_event(hess_matvec_buffer, event_order, forward_scratch_buffer)
+    Parameters
+    ----------
+    coxdev : CoxDeviance
+        The CoxDeviance object used for computations.
+    result : CoxDevianceResult
+        Result from the most recent deviance computation.
+        
+    Attributes
+    ----------
+    shape : tuple
+        Shape of the information matrix (n, n).
+    dtype : type
+        Data type of the matrix elements.
+    """
 
-    hess_matvec_buffer *= exp_w 
-    hess_matvec_buffer -= diag_part * arg
-    print(f'hess_matvec_buffer {hess_matvec_buffer}')
+    coxdev: CoxDeviance
+    result: CoxDevianceResult
+
+    def __post_init__(self):
+        """Initialize the linear operator dimensions."""
+        n = self.coxdev._status.shape[0]
+        self.shape = (n, n)
+        self.dtype = float
+        
+    def _matvec(self, arg):
+        """
+        Compute matrix-vector product with the information matrix.
+        
+        Parameters
+        ----------
+        arg : np.ndarray
+            Vector to multiply with the information matrix.
+            
+        Returns
+        -------
+        np.ndarray
+            Result of the matrix-vector multiplication.
+        """
+        # this will compute risk sums if not already computed
+        # at this linear_predictor and sample_weight
+        
+        result = self.result
+        coxdev = self.coxdev
+
+        # negative will give 2nd derivative of negative
+        # loglikelihood
+
+        _hessian_matvec(-np.asarray(arg).reshape(-1),
+                        np.asarray(result.linear_predictor),
+                        np.asarray(result.sample_weight),
+                        coxdev._risk_sum_buffers[0],
+                        coxdev._diag_part_buffer,
+                        coxdev._w_avg_buffer,
+                        coxdev._exp_w_buffer,
+                        coxdev._event_cumsum,
+                        coxdev._start_cumsum,
+                        coxdev._event_order,
+                        coxdev._start_order,
+                        coxdev._status,
+                        coxdev._first,
+                        coxdev._last,
+                        coxdev._scaling,
+                        coxdev._event_map,
+                        coxdev._start_map,
+                        coxdev._risk_sum_buffers,
+                        coxdev._forward_cumsum_buffers,
+                        coxdev._forward_scratch_buffer,
+                        coxdev._reverse_cumsum_buffers,
+                        coxdev._hess_matvec_buffer,
+                        coxdev._have_start_times,                        
+                        coxdev._efron)
+
+        return coxdev._hess_matvec_buffer.copy()
+
     
+    def _adjoint(self, arg):
+        """
+        Compute the adjoint (transpose) matrix-vector product.
+        
+        Since the information matrix is symmetric, this is the same as _matvec.
+        
+        Parameters
+        ----------
+        arg : np.ndarray
+            Vector to multiply with the adjoint matrix.
+            
+        Returns
+        -------
+        np.ndarray
+            Result of the adjoint matrix-vector multiplication.
+        """
+        # it is symmetric
+        return self._matvec(arg)
+
+
+# private functions
+
+def _preprocess(start,
+                event,
+                status):
+    """
+    Preprocess survival data for Cox model computations.
+    
+    This function handles data preprocessing including sorting, tie detection,
+    and creation of indexing arrays for efficient computation.
+    
+    Parameters
+    ----------
+    start : np.ndarray
+        Start times for left-truncated data.
+    event : np.ndarray
+        Event times.
+    status : np.ndarray
+        Event indicators.
+        
+    Returns
+    -------
+    tuple
+        Preprocessed data structures for efficient computation.
+    """
+    return c_preprocess(start, event, status)
