+++ tests/test_compareR.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pandas as pd
-from coxdev import CoxDeviance
+from coxdev import CoxDeviance, StratifiedCoxDeviance
 
 try:
     import rpy2.robjects as rpy
@@ -101,6 +101,80 @@ def get_coxph(event,
         D = rpy.r('F$loglik')
         cov = rpy.r('vcov(F)')
     return -2 * G, -2 * D, cov
+
+
+def get_stratified_coxph(event,
+                         status,
+                         strata,
+                         X,
+                         beta,
+                         sample_weight,
+                         start=None,
+                         ties='efron'):
+    """Get stratified Cox model results from R's survival package."""
+    
+    if start is not None:
+        start = np.asarray(start)
+    status = np.asarray(status)
+    event = np.asarray(event)
+    strata = np.asarray(strata)
+
+    with np_cv_rules.context():
+        rpy.r.assign('status', status)
+        rpy.r.assign('event', event)
+        rpy.r.assign('strata', strata)
+        rpy.r.assign('X', X)
+        rpy.r.assign('beta', beta)
+        rpy.r.assign('ties', ties)
+        rpy.r.assign('sample_weight', sample_weight)
+        rpy.r('sample_weight = as.numeric(sample_weight)')
+        rpy.r('strata = as.factor(strata)')
+        
+        if start is not None:
+            rpy.r.assign('start', start)
+            rpy.r('y = Surv(start, event, status)')
+        else:
+            rpy.r('y = Surv(event, status)')
+            
+        rpy.r('F = coxph(y ~ X + strata(strata), init=beta, weights=sample_weight, control=coxph.control(iter.max=0), ties=ties, robust=FALSE)')
+        rpy.r('score = colSums(coxph.detail(F)$scor)')
+        G = rpy.r('score')
+        D = rpy.r('F$loglik')
+        cov = rpy.r('vcov(F)')
+    return -2 * G, -2 * D, cov
+
+
+def create_stratified_data(n_samples=100, n_strata=3):
+    """Create stratified survival data for testing."""
+    np.random.seed(42)
+    
+    # Create strata
+    strata = np.random.randint(0, n_strata, n_samples)
+    
+    # Create survival data
+    event = np.random.exponential(1.0, n_samples)
+    status = np.random.binomial(1, 0.7, n_samples)
+    
+    # Add some start times for some tests
+    start = np.random.exponential(0.5, n_samples)
+    
+    # Create covariates
+    n_features = 3
+    X = np.random.standard_normal((n_samples, n_features))
+    beta = np.random.standard_normal(n_features) / np.sqrt(n_samples)
+    
+    # Create weights
+    weight = np.random.uniform(0.5, 2.0, n_samples)
+    
+    return {
+        'event': event,
+        'status': status,
+        'strata': strata,
+        'start': start,
+        'X': X,
+        'beta': beta,
+        'weight': weight
+    }
 
 
 @pytest.mark.parametrize('tie_types', all_combos)
@@ -213,3 +287,166 @@ def test_glmnet(tie_types,
 
     assert (delta_D < tol) and (delta_G < tol) and (delta_H < tol)
     
+
+@pytest.mark.parametrize('tie_breaking', ['efron', 'breslow'])
+@pytest.mark.parametrize('have_start_times', [True, False])
+@pytest.mark.parametrize('n_strata', [2, 3, 5])
+def test_stratified_coxph(tie_breaking, have_start_times, n_strata, tol=1e-10):
+    """Test StratifiedCoxDeviance against R's stratified coxph."""
+    
+    if not has_rpy2:
+        pytest.skip("rpy2 not available")
+    
+    # Create stratified data
+    data = create_stratified_data(n_samples=150, n_strata=n_strata)
+    
+    if have_start_times:
+        start = data['start']
+    else:
+        start = None
+    
+    # Create StratifiedCoxDeviance
+    stratdev = StratifiedCoxDeviance(
+        event=data['event'],
+        start=start,
+        status=data['status'],
+        strata=data['strata'],
+        tie_breaking=tie_breaking
+    )
+    
+    # Compute results with Python
+    eta = data['X'] @ data['beta']
+    C = stratdev(eta, data['weight'])
+    
+    # Get information matrix
+    H = stratdev.information(eta, data['weight'])
+    I = data['X'].T @ (H @ data['X'])
+    assert np.allclose(I, I.T)
+    cov_ = np.linalg.inv(I)
+    
+    # Get results from R
+    (G_coxph, D_coxph, cov_coxph) = get_stratified_coxph(
+        event=data['event'],
+        status=data['status'],
+        strata=data['strata'],
+        X=data['X'],
+        beta=data['beta'],
+        sample_weight=data['weight'],
+        start=start,
+        ties=tie_breaking
+    )
+    
+    # Compare deviance (adjust for saturated log-likelihood)
+    print(f"R deviance: {D_coxph[0]}, Python deviance: {C.deviance - 2 * C.loglik_sat}")
+    assert np.allclose(D_coxph[0], C.deviance - 2 * C.loglik_sat, rtol=tol)
+    
+    # Compare gradients
+    delta_ph = np.linalg.norm(G_coxph - data['X'].T @ C.gradient) / np.linalg.norm(data['X'].T @ C.gradient)
+    assert delta_ph < tol
+    
+    # Compare covariance matrices
+    assert np.linalg.norm(cov_ - cov_coxph) / np.linalg.norm(cov_) < tol
+
+
+@pytest.mark.parametrize('tie_breaking', ['efron', 'breslow'])
+@pytest.mark.parametrize('have_start_times', [True, False])
+def test_stratified_single_stratum(tie_breaking, have_start_times, tol=1e-10):
+    """Test that StratifiedCoxDeviance with single stratum matches CoxDeviance."""
+    
+    if not has_rpy2:
+        pytest.skip("rpy2 not available")
+    
+    # Create data with single stratum
+    data = create_stratified_data(n_samples=100, n_strata=1)
+    
+    if have_start_times:
+        start = data['start']
+    else:
+        start = None
+    
+    # Create both models
+    coxdev = CoxDeviance(
+        event=data['event'],
+        start=start,
+        status=data['status'],
+        tie_breaking=tie_breaking
+    )
+    
+    stratdev = StratifiedCoxDeviance(
+        event=data['event'],
+        start=start,
+        status=data['status'],
+        strata=data['strata'],
+        tie_breaking=tie_breaking
+    )
+    
+    # Compute results
+    eta = data['X'] @ data['beta']
+    C1 = coxdev(eta, data['weight'])
+    C2 = stratdev(eta, data['weight'])
+    
+    # Results should be identical
+    assert np.allclose(C1.deviance, C2.deviance, rtol=tol)
+    assert np.allclose(C1.gradient, C2.gradient, rtol=tol)
+    assert np.allclose(C1.diag_hessian, C2.diag_hessian, rtol=tol)
+    assert np.allclose(C1.loglik_sat, C2.loglik_sat, rtol=tol)
+
+
+@pytest.mark.parametrize('tie_breaking', ['efron', 'breslow'])
+def test_stratified_multiple_strata_sizes(tie_breaking, tol=1e-10):
+    """Test StratifiedCoxDeviance with varying stratum sizes."""
+    
+    if not has_rpy2:
+        pytest.skip("rpy2 not available")
+    
+    # Create data with uneven stratum sizes
+    np.random.seed(42)
+    n_samples = 200
+    
+    # Create strata with different sizes
+    strata = np.concatenate([
+        np.zeros(50),   # 50 samples in stratum 0
+        np.ones(80),    # 80 samples in stratum 1  
+        np.full(70, 2)  # 70 samples in stratum 2
+    ])
+    
+    # Create survival data
+    event = np.random.exponential(1.0, n_samples)
+    status = np.random.binomial(1, 0.7, n_samples)
+    start = np.random.exponential(0.5, n_samples)
+    
+    # Create covariates
+    n_features = 2
+    X = np.random.standard_normal((n_samples, n_features))
+    beta = np.random.standard_normal(n_features) / np.sqrt(n_samples)
+    weight = np.random.uniform(0.5, 2.0, n_samples)
+    
+    # Create StratifiedCoxDeviance
+    stratdev = StratifiedCoxDeviance(
+        event=event,
+        start=start,
+        status=status,
+        strata=strata,
+        tie_breaking=tie_breaking
+    )
+    
+    # Compute results
+    eta = X @ beta
+    C = stratdev(eta, weight)
+    
+    # Get results from R
+    (G_coxph, D_coxph, cov_coxph) = get_stratified_coxph(
+        event=event,
+        status=status,
+        strata=strata,
+        X=X,
+        beta=beta,
+        sample_weight=weight,
+        start=start,
+        ties=tie_breaking
+    )
+    
+    # Compare results
+    assert np.allclose(D_coxph[0], C.deviance - 2 * C.loglik_sat, rtol=tol)
+    delta_ph = np.linalg.norm(G_coxph - X.T @ C.gradient) / np.linalg.norm(X.T @ C.gradient)
+    assert delta_ph < tol
